+++ src/tmtccmd/config/definitions.py
@@ -68,19 +68,6 @@ ComIFDictT = Dict[str, ComIFValueT]
 EthernetAddressT = Tuple[str, int]
 
 
-class TmTypes(enum.Enum):
-    NONE = enum.auto()
-    CCSDS_SPACE_PACKETS = enum.auto()
-
-
-class TmHandler:
-    def __init__(self, tm_type: TmTypes):
-        self._tm_type = tm_type
-
-    def get_type(self):
-        return self._tm_type
-
-
 class CoreComInterfaces(enum.Enum):
     DUMMY = "dummy"
     SERIAL_DLE = "ser_dle"
