+++ src/tmtccmd/tm/definitions.py
@@ -1,3 +1,4 @@
+import enum
 from typing import Deque, Tuple, List
 from tmtccmd.ecss.tm import PusTelemetry
 
@@ -13,3 +14,8 @@ PusTmQueueT = Deque[PusTmListT]
 PusTmListT = List[PusTelemetry]
 PusTmObjQeue = Deque[PusTelemetry]
 PusTmTupleQueueT = Deque[PusTmTupleT]
+
+
+class TmTypes(enum.Enum):
+    NONE = enum.auto()
+    CCSDS_SPACE_PACKETS = enum.auto()
